+++ android/viewsets/westeros/src/main/java/com/liferay/mobile/screens/viewsets/westeros/ddl/form/DDLFormView.java
@@ -21,6 +21,7 @@ public class DDLFormView extends com.liferay.mobile.screens.viewsets.defaultview
 		WESTEROS_LAYOUT_IDS.put(Field.EditorType.TEXT, R.layout.ddlfield_text_westeros);
 		WESTEROS_LAYOUT_IDS.put(Field.EditorType.TEXT_AREA, R.layout.ddlfield_text_area_westeros);
 		WESTEROS_LAYOUT_IDS.put(Field.EditorType.DOCUMENT, R.layout.ddlfield_document_westeros);
+		WESTEROS_LAYOUT_IDS.put(Field.EditorType.GEO, R.layout.ddlfield_geo_default);
 	}
 
 	public DDLFormView(Context context) {